frontend/requirements.txt
streamlit>=1.30.0
streamlit-extras>=0.3.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
pyjwt>=2.10.1,<3.0.0

